+++ packages/common/src/Event.ts
@@ -1,6 +1,3 @@
-import { random } from './Utils';
 import { RpgCommonPlayer } from './Player'
 
-export default class Event extends RpgCommonPlayer {
-    
-}
+export class RpgEvent extends RpgCommonPlayer {}
