+++ tests/unit-tests/jest.config.js
@@ -1,8 +1,3 @@
 const jestConfig = require('@rpgjs/compiler/jest')
 
-jestConfig.transformIgnorePatterns = [
-    ...jestConfig.transformIgnorePatterns,
-    ...['[/\]core-js', '@babel[/\]runtime']
-]
-
 module.exports = jestConfig
